Refactor: RandomGenerator 생성하여 RectangleFactory, plane 중복코드 제거
