+++ DrawingApp/DrawingApp/Model/RectangleFactory.swift
@@ -8,35 +8,15 @@
 import Foundation
 
 struct RectangleFactory {
-//    private let screenSize : (Double,Double)
 
     static func makeRandomRectangle(in screenSize : (Double,Double)) -> Rectangle? {
-        var randomPosition : Position {
-            let screenWidth = screenSize.0
-            let screenHeight = screenSize.1
-            let randomCoordinate = {range in
-                Double.random(in: 0...range)
-            }
-            let randomX = randomCoordinate(screenWidth)
-            let randomY = randomCoordinate(screenHeight)
-            return Position(x: randomX, y: randomY)
-        }
+        let randomPosition = RandomGenerator.generatePosition(screenSize: screenSize)
+        let randomColor = RandomGenerator.generateColor()
+        let randomAlpha = RandomGenerator.generateAlpha()
         
-        var randomColor : Color? {
-            var randomInt: Int {
-                Color.colorRange.randomElement() ?? 0
-            }
-            guard let randomColor = Color(red: randomInt, green: randomInt, blue: randomInt) else {return nil}
-            return randomColor
-        }
-
-        
-        var randomAlpha : Alpha? {
-            let randomTransparency = Double.random(in: Alpha.alphaRange)
-            return Alpha(transparency: randomTransparency)
-        }
         guard let randomColor = randomColor else {return nil}
         guard let randomAlpha = randomAlpha else {return nil}
+        
         return Rectangle(size: Size(width: 150, height: 120), position: randomPosition, color: randomColor, alpha: randomAlpha)
     }
